[17/n][Services] Tunnelbroker - Config manager.
Summary: We currently have hard-coded config in Constants.h. This diff moves some of that config into tunnelbroker.env, which is passed into the container environment by docker-compose, and then loads the config into a struct that can be accessed as const.

Test Plan: I tested that all the environment variables are successfully loaded into the TunnelbrokerConfig struct if provided, and that the program exits if the environment variables aren't set.

Reviewers: geekbrother, karol-bisztyga, palys-swm, ashoat

Reviewed By: ashoat

Subscribers: benschac, ashoat, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D2903